pyproject.toml
[project]
name = "logs-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "gentrace-py>=0.42.0",
    "openai>=1.65.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]

[tool.uv.workspace]
members = ["evalx"]
